#46 Move new Models dedicated namespace
